feat: split the main modules into multiple modules - DIA-61984
