+++ fastapi_sqla/__init__.py
@@ -1,17 +1,7 @@
-import os
-
-from fastapi import FastAPI
-
-from fastapi_sqla import sqla
-from fastapi_sqla.models import Collection, Item, Page
-from fastapi_sqla.sqla import (
-    Base,
-    Paginate,
-    PaginateSignature,
-    Pagination,
-    Session,
-    open_session,
-)
+from fastapi_sqla.base import setup
+from fastapi_sqla.models import Base, Collection, Item, Page
+from fastapi_sqla.pagination import Paginate, PaginateSignature, Pagination
+from fastapi_sqla.sqla import Session, open_session
 
 __all__ = [
     "Base",
@@ -23,17 +13,14 @@ __all__ = [
     "Pagination",
     "Session",
     "open_session",
+    "setup",
 ]
 
 
 try:
-    from fastapi_sqla import asyncio_support
-    from fastapi_sqla.asyncio_support import (  # noqa
-        AsyncPaginate,
-        AsyncPagination,
-        AsyncSession,
-    )
-    from fastapi_sqla.asyncio_support import open_session as open_async_session  # noqa
+    from fastapi_sqla.async_pagination import AsyncPaginate, AsyncPagination
+    from fastapi_sqla.asyncio_support import AsyncSession
+    from fastapi_sqla.asyncio_support import open_session as open_async_session
 
     __all__ += [
         "AsyncPaginate",
@@ -43,22 +30,5 @@ try:
     ]
     has_asyncio_support = True
 
-except ImportError as err:  # pragma: no cover
-    has_asyncio_support = False
-    asyncio_support_err = str(err)
-
-
-def setup(app: FastAPI):
-    engine = sqla.new_engine()
-
-    if not sqla.is_async_dialect(engine):
-        app.add_event_handler("startup", sqla.startup)
-        app.middleware("http")(sqla.add_session_to_request)
-
-    has_async_config = "async_sqlalchemy_url" in os.environ or sqla.is_async_dialect(
-        engine
-    )
-    if has_async_config:
-        assert has_asyncio_support, asyncio_support_err
-        app.add_event_handler("startup", asyncio_support.startup)
-        app.middleware("http")(asyncio_support.add_session_to_request)
+except ImportError:  # pragma: no cover
+    pass
