+++ fastapi_sqla/asyncio_support.py
@@ -1,20 +1,18 @@
-import math
 import os
-from collections.abc import AsyncGenerator, Awaitable, Callable
+from collections.abc import AsyncGenerator
 from contextlib import asynccontextmanager
-from typing import Iterator, Optional, Union, cast
+from typing import cast
 
 import structlog
-from fastapi import Depends, Query, Request
+from fastapi import Request
 from sqlalchemy import text
 from sqlalchemy.ext.asyncio import AsyncEngine
 from sqlalchemy.ext.asyncio import AsyncSession as SqlaAsyncSession
 from sqlalchemy.orm.session import sessionmaker
-from sqlalchemy.sql import Select, func, select
 
 from fastapi_sqla import aws_aurora_support, aws_rds_iam_support
-from fastapi_sqla.models import Page
-from fastapi_sqla.sqla import Base, new_engine
+from fastapi_sqla.models import Base
+from fastapi_sqla.sqla import new_engine
 
 logger = structlog.get_logger(__name__)
 _ASYNC_SESSION_KEY = "fastapi_sqla_async_session"
@@ -127,82 +125,3 @@ async def add_session_to_request(request: Request, call_next):
             await session.rollback()
 
     return response
-
-
-QueryCountDependency = Callable[..., Awaitable[int]]
-PaginateSignature = Callable[[Select, Optional[bool]], Awaitable[Page]]
-DefaultDependency = Callable[[AsyncSession, int, int], PaginateSignature]
-WithQueryCountDependency = Callable[[AsyncSession, int, int, int], PaginateSignature]
-PaginateDependency = Union[DefaultDependency, WithQueryCountDependency]
-
-
-async def default_query_count(session: AsyncSession, query: Select) -> int:
-    result = await session.execute(select(func.count()).select_from(query.subquery()))
-    return cast(int, result.scalar())
-
-
-async def paginate_query(
-    query: Select,
-    session: AsyncSession,
-    total_items: int,
-    offset: int,
-    limit: int,
-    *,
-    scalars: bool = True,
-) -> Page:
-    total_pages = math.ceil(total_items / limit)
-    page_number = offset / limit + 1
-    query = query.offset(offset).limit(limit)
-    result = await session.execute(query)
-    data = iter(
-        cast(Iterator, result.unique().scalars() if scalars else result.mappings())
-    )
-    return Page(
-        data=data,
-        meta={
-            "offset": offset,
-            "total_items": total_items,
-            "total_pages": total_pages,
-            "page_number": page_number,
-        },
-    )
-
-
-def AsyncPagination(
-    min_page_size: int = 10,
-    max_page_size: int = 100,
-    query_count: Union[QueryCountDependency, None] = None,
-) -> PaginateDependency:
-    def default_dependency(
-        session: AsyncSession = Depends(),
-        offset: int = Query(0, ge=0),
-        limit: int = Query(min_page_size, ge=1, le=max_page_size),
-    ) -> PaginateSignature:
-        async def paginate(query: Select, scalars=True) -> Page:
-            total_items = await default_query_count(session, query)
-            return await paginate_query(
-                query, session, total_items, offset, limit, scalars=scalars
-            )
-
-        return paginate
-
-    def with_query_count_dependency(
-        session: AsyncSession = Depends(),
-        offset: int = Query(0, ge=0),
-        limit: int = Query(min_page_size, ge=1, le=max_page_size),
-        total_items: int = Depends(query_count),
-    ):
-        async def paginate(query: Select, scalars=True) -> Page:
-            return await paginate_query(
-                query, session, total_items, offset, limit, scalars=scalars
-            )
-
-        return paginate
-
-    if query_count:
-        return with_query_count_dependency
-    else:
-        return default_dependency
-
-
-AsyncPaginate: PaginateDependency = AsyncPagination()
