+++ fastapi_sqla/models.py
@@ -2,6 +2,14 @@ from typing import Generic, TypeVar
 
 from pydantic import BaseModel, Field
 from pydantic import __version__ as pydantic_version
+from sqlalchemy.ext.declarative import DeferredReflection
+
+try:
+    from sqlalchemy.orm import DeclarativeBase
+except ImportError:
+    from sqlalchemy.ext.declarative import declarative_base
+
+    DeclarativeBase = declarative_base()  # type: ignore
 
 major, _, _ = [int(v) for v in pydantic_version.split(".")]
 is_pydantic2 = major == 2
@@ -10,19 +18,24 @@ if is_pydantic2:
 else:
     from pydantic.generics import GenericModel  # type:ignore
 
-T = TypeVar("T")
+
+class Base(DeclarativeBase, DeferredReflection):
+    __abstract__ = True
 
 
-class Item(GenericModel, Generic[T]):
+ItemT = TypeVar("ItemT")
+
+
+class Item(GenericModel, Generic[ItemT]):
     """Item container."""
 
-    data: T
+    data: ItemT
 
 
-class Collection(GenericModel, Generic[T]):
+class Collection(GenericModel, Generic[ItemT]):
     """Collection container."""
 
-    data: list[T]
+    data: list[ItemT]
 
 
 class Meta(BaseModel):
@@ -34,7 +47,7 @@ class Meta(BaseModel):
     page_number: int = Field(..., description="Current page number. Starts at 1.")
 
 
-class Page(Collection[T], Generic[T]):
+class Page(Collection[ItemT], Generic[ItemT]):
     """A page of the collection with info on current page and total items in meta."""
 
     meta: Meta
